requirements.txt
# Core FastAPI & Web Framework
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
python-multipart==0.0.9

# LLM & AI Libraries
openai==1.12.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
transformers==4.37.2
torch==2.1.2
sentence-transformers==2.2.2

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3

# Vector Database & Search
redis==5.0.1
psycopg2-binary==2.9.9
pgvector==0.2.4
chromadb==0.4.22

# Policy Engine
opa==0.1.0

# Monitoring & Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
prometheus-client==0.19.0

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
celery==5.3.4
redis==5.0.1
pandas==2.1.4
openpyxl==3.1.2
reportlab==4.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.26.0

# Development
black==23.12.1
isort==5.13.2
flake8==6.1.0
mypy==1.8.0
